+++ src/boolean.ts
@@ -1,8 +1,8 @@
 const falsyStrings = ["n", "no", "f", "false"];
 
 /**
- * Returns true if some string can be interpreted as false.
+ * Returns true if some string can be interpreted as true.
  */
-export function isFalsy(value: string): boolean {
+export function isTruthy(value: string): boolean {
 	return !falsyStrings.includes(value.trim().toLowerCase());
 }